Fixed a bug with retriggering Skeletal Mage Fire Nova ability
This feels like a bug because we need to check the bool on the ability in order to retrigger it because it is already triggered for some reason and it fails in InternalTryToActivateAbility().
I don't really know why there is already already one activated. It only happens when there are two abilities used by Skeletal Mage in its BT