Refactor item pickup logic for out-of-range items to work on client and fix a item dupe bug
Extracted and centralized logic for handling item pickups when the item is out of range into a new HandlePickUpIfItemOutOfRange function. Added a client RPC (ClientStartApproachingOutOfRangeItem) to initiate auto-run towards the item, and updated ServerPickupItem_Implementation and ServerGrabDroppableItemToCursor_Implementation to use the new logic. Introduced EObsidianItemPickUpType enum for pickup type distinction.